Got compiling of 32-bit test cases working. The issue was related to scalarization of trivial structures containing a reference.
Added more constraints on the packing of the State (sub)structures.

Trying to figure out why the x86 tests fail. Issue seems to be that they are operating on the FPU IP at the machine code level, but appear to be correctly operating on edi/esi at the bitcode level.