Implement the majority of the subprocess code to work with radare2. This may change: currently it attempts to gather all function disassemblies in a single whack, which could be memory problematic. Unfortunatley the way I have things, it is passing a script to radare2 to do the commands I want. I did this because of the issues with subprocess and semi-colons, which I had issues working with escaping and still having them work in radare2. This perhaps means that when/if the code changes to analyzing function by function, that a Popen version should be implemented.
Further, there is also r2pipe module which I have not tried. It is
likely dealing with the functionality I want, but I initially did
not want to rely on that being installed. :-/.

I think I will leave this as-is, for now, work on mapping my CFG
stuffer to this tool interface, and then re-do the r2 io.