Add X87 FBLD, FBSTP (#219)
* Add types, read operator for packed BCD

* Add FBLD semantic

* Improve field naming, internal structure of `dec80_t`

Rename `digits` to `digit_pairs` to make it obvious that there are two
decimal digits encoded at each index. This change and the bitfield view
provided by the union will support more readable code when we implement
the MD80 write operator and FBSTP semantic.

* Add write operators for packed BCD, BCD indefinite

* Add FBSTP semantic

* Factor out `dec_digit_pair_t`, rename `pair` fields

* Improve BCD range checking in FBSTP

- Do bounds check on rounded value
- Check again empirically-derived value that accounts for aliasing
- Add more test inputs and improve comments

* Change naming of packed BCD types and operators

The Intel SDM is inconsistent in its naming of operand types, sometimes
using "m80dec" and sometimes "m80bcd". "BCD" is used most frequently in
the actual text, is very specific, and avoids any confusion with base 10
floating-point formats, so we switch to that.

* Don't use typedef for `bcd_digit_pair_t`