Ubuntu 18.04 Dockerfile to produce compatible linkable libraries on Nessie
To create the compatible binaries:
```
$ docker build -f Dockerfile.vcpkg --target base -t vcpkg-base-18.04 .

$ docker run --rm -t -i -u $(id -u ${USER}):$(id -g ${USER}) \
    -v $(pwd):/workspace -w /workspace vcpkg-base-18.04 \
    ./vcpkg/vcpkg install --overlay-ports=./vcpkg-overlays \
    glog gflags xed gtest
```

To use the libraries, perform the necessary OS checks to make sure that
the Dockerfile image used to create the libraries is the same Linux
distro you are using. The built package hashes _should_ match if you are
using the same compiler executable (same hash) and you are building from
the same checkout of vcpkg.

If you are having issues, run vcpkg with --debug to get more output.

To have vcpkg look for (and eventually automatically download) the
cached binaries, execute the following on your host machine:

```
$ export CC=$(which clang-10)
$ export CXX=$(which clang++-10)
$ export CPP=$(which clang++-10)

$ ./vcpkg/bootstrap-vcpkg.sh
$ ./vcpkg/vcpkg install glog gflags xed gtest

$ mkdir build && cd build
$ cmake -G Ninja -DCMAKE_TOOLCHAIN_FILE=../vcpkg/scripts/buildsystems/vcpkg.cmake \
    -DLLVM_DIR=/usr/lib/llvm-8/cmake -DCMAKE_INSTALL_PREFIX=$(pwd)/install ..
$ cmake --build .
$ cmake --build . --target install
$ cmake --build . --target test_dependencies
$ env CTEST_OUTPUT_ON_FAILURE=1 cmake --build . --target test
```

Note: LLVM from vcpkg hasn't been tested, so we're using LLVM-8 that was
locally installed on Ubuntu 18.04.